Clean up osu object inspector
