+++ PerformanceCalculatorGUI/Screens/ObjectInspection/OsuObjectInspectorDrawable.cs
@@ -1,14 +1,17 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Performance;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Utils;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Pooling;
+using osu.Game.Rulesets.Osu.Difficulty.Preprocessing;
 using osu.Game.Rulesets.Osu.Edit;
 using osu.Game.Rulesets.Osu.Objects;
 
@@ -106,6 +109,58 @@ namespace PerformanceCalculatorGUI.Screens.ObjectInspection
             }
 
             entry.LifetimeEnd = panel.LifetimeEnd;
+        }
+    }
+
+    internal class OsuObjectInspectorLifetimeEntry : LifetimeEntry
+    {
+        public event Action? Invalidated;
+        public readonly OsuHitObject HitObject;
+        public readonly OsuDifficultyHitObject DifficultyHitObject;
+
+        public OsuObjectInspectorLifetimeEntry(OsuHitObject hitObject, OsuDifficultyHitObject difficultyHitObject)
+        {
+            HitObject = hitObject;
+            DifficultyHitObject = difficultyHitObject;
+            LifetimeStart = HitObject.StartTime - HitObject.TimePreempt;
+
+            bindEvents();
+            refreshLifetimes();
+        }
+
+        private bool wasBound;
+
+        private void bindEvents()
+        {
+            UnbindEvents();
+            HitObject.DefaultsApplied += onDefaultsApplied;
+            wasBound = true;
+        }
+
+        public void UnbindEvents()
+        {
+            if (!wasBound)
+                return;
+
+            HitObject.DefaultsApplied -= onDefaultsApplied;
+
+            wasBound = false;
+        }
+
+        private void onDefaultsApplied(HitObject obj) => refreshLifetimes();
+
+        private void refreshLifetimes()
+        {
+            if (HitObject is Spinner)
+            {
+                LifetimeEnd = LifetimeStart;
+                return;
+            }
+
+            LifetimeStart = HitObject.StartTime - HitObject.TimePreempt;
+            LifetimeEnd = HitObject.GetEndTime() + DrawableOsuEditorRuleset.EDITOR_HIT_OBJECT_FADE_OUT_EXTENSION;
+
+            Invalidated?.Invoke();
         }
     }
 }
