+++ PerformanceCalculatorGUI/Screens/ObjectInspection/OsuObjectInspectorRenderer.cs
@@ -3,10 +3,8 @@
 
 using System.Collections.Generic;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Pooling;
-using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Pooling;
 using osu.Game.Rulesets.Osu.Difficulty.Preprocessing;
 using osu.Game.Rulesets.Osu.Objects;
@@ -18,7 +16,6 @@ namespace PerformanceCalculatorGUI.Screens.ObjectInspection
         private DrawablePool<OsuObjectInspectorDrawable> pool;
 
         private readonly List<OsuObjectInspectorLifetimeEntry> lifetimeEntries = new();
-        private readonly Dictionary<HitObject, IBindable> startTimeMap = new();
 
         [BackgroundDependencyLoader]
         private void load()
@@ -31,29 +28,12 @@ namespace PerformanceCalculatorGUI.Screens.ObjectInspection
 
         public void AddDifficultyDataPanel(OsuHitObject hitObject, OsuDifficultyHitObject difficultyHitObject)
         {
-            addEntry(hitObject, difficultyHitObject);
-
-            var startTimeBindable = hitObject.StartTimeBindable.GetBoundCopy();
-            startTimeBindable.ValueChanged += _ => onStartTimeChanged(hitObject, difficultyHitObject);
-            startTimeMap[hitObject] = startTimeBindable;
-        }
-
-        public void RemoveDifficultyDataPanel(OsuHitObject hitObject)
-        {
-            removeEntry(hitObject);
-
-            startTimeMap[hitObject].UnbindAll();
-            startTimeMap.Remove(hitObject);
-        }
-
-        private void addEntry(OsuHitObject hitObject, OsuDifficultyHitObject difficultyHitObject)
-        {
             var newEntry = new OsuObjectInspectorLifetimeEntry(hitObject, difficultyHitObject);
             lifetimeEntries.Add(newEntry);
             Add(newEntry);
         }
 
-        private void removeEntry(OsuHitObject hitObject)
+        public void RemoveDifficultyDataPanel(OsuHitObject hitObject)
         {
             int index = lifetimeEntries.FindIndex(e => e.HitObject == hitObject);
 
@@ -69,12 +49,6 @@ namespace PerformanceCalculatorGUI.Screens.ObjectInspection
             var connection = pool.Get();
             connection.Apply(entry);
             return connection;
-        }
-
-        private void onStartTimeChanged(OsuHitObject hitObject, OsuDifficultyHitObject difficultyHitObject)
-        {
-            removeEntry(hitObject);
-            addEntry(hitObject, difficultyHitObject);
         }
 
         protected override void Dispose(bool isDisposing)
